requirements.txt
gradio
gradio[oauth]
requests
langchain
langchain-community
langchain-core
langchain-google-genai
langchain-huggingface
langchain-ollama
langchain-together
langchain-tavily
langchain-chroma
langgraph
huggingface_hub
supabase
arxiv
pymupdf
wikipedia
pgvector
python-dotenv
ipykernel
ddgs
wikipedia
wiki
bs4
openpyxl
pillow